added computer save method
