+++ LabManager/Repositories/ComputerRepository.cs
@@ -2,7 +2,6 @@ using LabManager.Database;
 using LabManager.Models;
 using Microsoft.Data.Sqlite;
 namespace LabManager.Repositories;
-
 class ComputerRepository
 {
     private readonly DatabaseConfig _databaseConfig;
@@ -16,7 +15,6 @@ class ComputerRepository
         var computers = new List<Computer>();
         var connection = new SqliteConnection(_databaseConfig.ConnectionString);
         connection.Open();
-
         var command = connection.CreateCommand();
         command.CommandText = "SELECT * FROM Computers";
         var reader = command.ExecuteReader();
@@ -31,5 +29,22 @@ class ComputerRepository
         connection.Close();
 
         return computers;
+    }
+
+    public Computer Save(Computer computer)
+    {
+        var connection = new SqliteConnection("Data Source=database.db");
+        connection.Open();
+
+        var command = connection.CreateCommand();
+        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor)";
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Processor);
+
+        command.ExecuteNonQuery();
+        connection.Close();
+
+        return computer;
     }
 }
